requirements.txt
altgraph==0.17.4
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
numpy==2.3.0
packaging==25.0
pandas==2.3.3
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.9
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==6.2.0
requests==2.32.4
setuptools==80.9.0
six==1.17.0
spotipy==2.25.1
tzdata==2025.2
urllib3==2.5.0
